refactor: 타입 정의를 각 api 파일로 이동
