+++ src/apis/vote.ts
@@ -1,4 +1,15 @@
-import { VoteResult } from '@/src/apis/types';
+interface Option {
+  seq: number;
+  content: string;
+  count: number;
+}
+
+export interface VoteResult {
+  voteId: number;
+  title: string;
+  totalNumber: number;
+  options: Option[];
+}
 
 export const getVoteResult = async (voteId: string) => {
   const data = (
